--- tests/api/test_show.py
@@ -1,9 +1,0 @@
-from gto.api import show
-from tests.api import data
-from tests.skip_presets import skip_for_windows_py_lt_3_9
-
-
-@skip_for_windows_py_lt_3_9
-def test_if_show_on_remote_git_repo_then_return_expected_registry():
-    result = show(repo=data.SAMPLE_REMOTE_REPO_URL)
-    assert result == data.get_sample_remote_repo_expected_registry()
