--- tests/git_utils/data/__init__.py
@@ -1,4 +1,0 @@
-SAMPLE_HTTP_REMOTE_REPO = "https://github.com/iterative/example-gto.git"
-SAMPLE_HTTP_REMOTE_REPO_WITHOUT_DOT_GIT_SUFFIX = (
-    "https://github.com/iterative/example-gto"
-)
