+++ tests/resources/sample_remote_repo_expected_registry.json
